rubiks-cube-web/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
opencv-python==4.9.0.80
numpy==1.26.3
Pillow==10.2.0
python-dotenv==1.0.0
gunicorn==21.2.0